Run DEGs, switching on the method

Note: It is recommended to save subset information in metadata(se)$subset since there is technically no way to recover that. However, from sample_metadata this should be inferrable.

deg(
  se,
  pathology,
  case,
  control,
  covariates,
  method,
  output = NULL,
  sample.col = "Sample",
  cpm.cutoff = 10,
  filter_only_case_control = TRUE,
  NRUV = 0,
  min.total.counts.per.sample = 100,
  IQR.factor = 1.5,
  outlier.covariates = c("log1p_total_counts", "n_genes_by_counts", "pct_counts_mt",
    "pct_counts_ribo"),
  verbose = TRUE,
  ensure.integer.counts = TRUE,
  mc.cores = getOption("mc.cores", 12)
)

Arguments

se

SummarizedExperiment object, with integer counts in the "counts" assay.

pathology

Column in colData(se) corresponding to diagnosis or experimental condition

case

Column in colData(se) corresponding to case values, which must be in "pathology"

control

Column in colData(se) corresponding to control values in "pathology"

covariates

Covariates to be used. Non-existent variables will be removed, and RUV variables will be added (if NRUV>0).

method

Method(s) of DEGs to be used. Valid values are DESeq2, edgerR-GLM, edgeR-QL, nebula, mast-hurdle, mast-random-effect

output

Output XLSX

sample.col

Sample column in colData(se) corresponding to the experiment. This is the column on which data is pseudobulked.

cpm.cutoff

Counts per Million cutoff used for filtering lowly-expressed genes. Genes must be expressed in at least min(case, control) samples at least at this log(CPM) level to be used. This n=min(case, control) is after outlier removal and low-count data.

filter_only_case_control

Filter se to only case+control only. Default TRUE

NRUV

Number of RUVSeq components to include

IQR.factor

Number of IQR factors above 75% percentile to be removed, according to scaled PCA of outlier.covariates. Set to Inf to remove none.

outlier.covariates

Sample-level covariates which are scaled and PCA transformed, to get distances from the centroid of the PCA embedding. Outliers above 75% + IQR.factor * IQR are removed.